feat: 🎸 finish the department class
✅ Closes: #5
